Core/DBLayer: Optimized prepared statement query results by eliminating unneeded buffer copies
* Improved error logs for using incorrect Field getters to also include table name, field name and field index.

+++ src/server/game/Globals/ObjectMgr.cpp
@@ -5586,7 +5586,7 @@ void ObjectMgr::LoadQuestGreetings()
                 continue;
         }
 
-        uint16 greetEmoteType       = fields[2].GetUInt32();
+        uint16 greetEmoteType       = fields[2].GetUInt16();
         uint32 greetEmoteDelay      = fields[3].GetUInt32();
         std::string greeting        = fields[4].GetString();
